Update templates/bump version number.
THIS CHANGE IS BREAKING, PLEASE UPDATE CONFIGS.

Decided not to go with a more robust config as we are using environment variables for production, we can be scripted before launching the app. I think it's good as-is for now, until the driver gets removed from db-migrate-mysql.

Version upgrades for packages.