Remove unused import in backupapi
Fixed demoapi to rename the demo to zip as we are zipping the file, just not renaming the extension.

Update veto picks in maplflow service to choose the correct maps from the in-game veto pool.

Move `updatePlayerStats` to the Utils class as it's used in seriesflow and mapflow services.

Update seriesflowservice to update the stats a final time to ensure data correctness at the end of a map.